Run dependencies only once and in the correct order
Adds that dependencies, which before this change could run multiple times, concurrently, are now still run concurrently, but only once, and in the correct order.

Example:
a depends on b and c
b depends on c
c depends on d
d depends on nothing

When getting the dependencies of a, the following is found:
[0]{"b","c"} - because a depends on b and c
[1]{"c","d"} - because b (from level 0) depends on c, and c (from level 0) depends on d
[2]{"d"} - because c (from level 1) depends on d

Knowing now the dependency tiers, we can start running the dependencies at the deepest level (2) and move up until the highest level (0).

If we see that a dependency has already run, we don't need to run it again. Dependencies still run concurrently per level.