Include pre, post, and final run functions
This allows the user to customize the run, wrapping the output of each task run
with an execution of a pre and post function.  A deferred final function is also
included to ensure any cleanup, if necessary.

This is being used as a library, where task is injected from another process.