Simplify vectorial logo paths
Previous version of the logo had complex shapes which caused rendering
issues in some cases (a thin horizontal line appearing). This change
ensures only 3 simple shapes are used and sorts the visual artifacts.

A side effect is also a small decrease in size. Apparently we do not
need to update the raster logos as they do not present the bug.