refactor: Re-organize node loading code to make it easier to follow
Readability is subjective, but while working in this part of the code
I found it difficult to follow the different code paths when loading
a node's content.

In this change I am attempting to make the branches of the decision
tree return early as much as possible.  This way we discard each
possibility as we work our way down, in this order:

1. Non-remote nodes
2. In offline mode, use cached remote nodes
3. On network timeout, use cached remote nodes
4a. Use fetched remote node
4b. On checksum change, handle writing to cache

Tests are added to ensure this functionality does not break.