feat: trigger watched task on new file
Currently, if a task is being watched and a file matching the source glob gets
added, the task is NOT run; the file needs to be *subsequently* modified to
trigger an actual task run. This PR changes this behavior so that the simple act
of adding a file will trigger a task run.