docs: update link to package.nix @ Nixpkgs
Update reference to the branch in the link (nixos-unstable → master),
to be consistent with the previous explanation.

To clarify, the updates to the packages in Nixpkgs are performed on the
master branch, and later migrated to the different production
branches (nixpkgs-unstable, nixos-unstable, nixos-unstable-small), and
possibly backported to the stable branches.