docs: update link to Nixpkgs’ source file
Earlier this year Nixpkgs directory structure was refactored to a
hashed by-name scheme.  Other links in the doc were already updated,
this one was still pending.

Also, I’ve changed the label from NixOS to Nix to reflect the wider
scope.  That is, the go-task package at Nixpkgs can be instaled via the
Nix package manager wherever you have Nix installed, be it NixOS (or
virtually any other Linux distribution, for that matter), or even macOS.